identity_matching/requirements.txt
numpy
pandas
scikit-learn
pyyaml
torch
clip-by-openai
opencv-python
Pillow
tqdm